renamed NLFile header file to a more appropriate name as functions.h
