+++ functions.h
